fix LDAP server failover by implementing our own failover mechanism instead of relying on the LDAP SDK's.
ldap server failover (space-separated list of hostnames in the argument to
LDAPServer) has been broken since 2.8.19 when ldap_init() was replaced with
ldap_initialize(), if available. as of that commit, a LDAP URL is
constructed if simple hostname(s) are passed to LDAPServer, so:

  'LDAPServer "127.0.0.1 172.16.0.5"

was passed to ldap_initailize() as:

  ldap://127.0.0.1 172.16.0.5/

Unfortunately, ldap_url_parse() treats this as a valid URL, but
ldap_intiailize() will not fail over to subsequent LDAP servers
should the first be unreachable.

ldap_initialize() *will* take a whitespace-separated list of LDAP URLs and
will fail over successfully, just not when whitespace-separated hostname
are passed to a *single* url (even if the whitespace in the hostname portion
of the URL is urlencoded as %20).

forget all of this SDK-based failover and implement our own failover support,
since afaict Sun's ancient LDAP SDK doesn't implement failover and it's nice
to be able to control it ourselves.

for that matter, we *need* to control it ourselves since we need to use the
search scope indicated by the URL at hand. chances are, the search scope
will be the same for all URLs passed to LDAPServer, but to be completely
correct, we should use the scope corresponding to the current URL and
ldap_initialize() provides no way to tell what URL it wound up connecting
to.

[oops, fired off the editor prematurely on the last commit.]