thanks to Bug 2922 (Auth API allows one auth module to authenticate user data provided by a different auth module), we don't need to worry about passing a locked crypted password ("*") to pr_auth_check(), since our authz handler will only be called if we supplied the original user information.
as a result, if the UserPassword directive is in effect, we need to avoid
doing auth binds and fall through to verifying the crypted password
ourselves instead of declining the authz request and letting mod_auth_unix
handle it.

bump the required proftpd version to 1.3.1rc3, since #2922 was resolved in
that version. this seems reasonable, since nobody ought to be running
1.2.10rc2 any more and 1.3.1rc3 has been available for almost 2.5y now.