remove LDAPQuotas in favor of using the LDAPUsers configuration
don't see a point in configuring quota lookups separately, since the quota
attrs should be on the user entry itself. if ldap quota lookups should be
disabled, "ldap:" can be removed QuotaLimitTable.

this seems to cut down on duplicate configuration and hopefully continues
the trend of making mod_ldap easier to configure.